requirements.txt
# Core Framework
python-telegram-bot[all]==20.7

# AI Engine
google-generativeai==0.8.3

# Environment & Utilities
python-dotenv==1.0.1